Refactored SingleUserPage and related folders
