+++ src/pages/SingleUserPage/components/Posts/Posts.tsx
@@ -4,7 +4,7 @@ import Typography from "antd/es/typography";
 import Empty from "antd/es/empty";
 
 import { Post as UserPost } from "./components";
-import useFetchPosts from "../../shared/hooks/useFetchPosts";
+import { useFetchPosts } from "./hooks/";
 import type { Post } from "./types/index";
 
 const Posts = () => {
@@ -15,9 +15,7 @@ const Posts = () => {
     <Flex vertical flex={1} gap={10} justify="center" align="center">
       <Typography.Title level={4}>User Post's</Typography.Title>
       {posts.length > 0 ? (
-        posts.map((post: Post) => (
-          <UserPost key={post.id} post={post} />
-        ))
+        posts.map((post: Post) => <UserPost key={post.id} post={post} />)
       ) : (
         <Empty />
       )}
