+++ src/pages/SingleUserPage/components/Posts/components/Post/components/index.ts
@@ -1,1 +1,3 @@
 export { default as ConfirmationModal } from './ConfirmationModal/ConfirmationModal'
+export { default as Actions } from './Actions/Actions'
+export { default as EditActions } from './EditActions/EditActions'
