+++ src/pages/SingleUserPage/components/Posts/components/Post/hooks/index.ts
@@ -1,0 +1,2 @@
+export {default as useDeletePost} from './useDeletePost';
+export {default as useUpdatePost} from './useUpdatePost'
