+++ src/pages/SingleUserPage/components/Posts/components/index.ts
@@ -1,3 +1,1 @@
 export { default as Post } from "./Post/Post";
-export {default as EditActions} from './EditActions/EditActions'
-export {default as Actions} from './Actions/Actions'
