+++ src/pages/SingleUserPage/components/Posts/hooks/index.ts
@@ -1,0 +1,1 @@
+export {default as useFetchPosts} from './useFetchPosts'
